requirements.txt
# Core dependencies
requests>=2.28.0
aiohttp>=3.8.0
pystray>=0.19.4
plyer>=2.1.0
schedule>=1.2.0
keyring>=24.0.0
pyyaml>=6.0
click>=8.1.0

# GUI dependencies
Pillow>=9.0.0

# Encryption
cryptography>=3.4.8

# Development dependencies  
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Packaging
pyinstaller>=5.0.0